+++ scilab/modules/graphics/src/c/getHandleProperty/getHandleProperty.h
@@ -166,6 +166,7 @@ GRAPHICS_IMPEXP void* get_UID(void* _pvCtx, int iObjUID);
 GRAPHICS_IMPEXP void* get_tip_orientation_property(void* _pvCtx, int iObjUID);
 GRAPHICS_IMPEXP void* get_tip_z_component_property(void* _pvCtx, int iObjUID);
 GRAPHICS_IMPEXP void* get_tip_display_components_property(void* _pvCtx, int iObjUID);
+GRAPHICS_IMPEXP void* get_datatip_display_mode_property(void* _pvCtx, int iObjUID);
 GRAPHICS_IMPEXP void* get_tip_auto_orientation_property(void* _pvCtx, int iObjUID);
 GRAPHICS_IMPEXP void* get_tip_interp_mode_property(void* _pvCtx, int iObjUID);
 GRAPHICS_IMPEXP void* get_tip_box_mode_property(void* _pvCtx, int iObjUID);